web: move project setup surfaces into header sheets
Relocate the three setup cards that sat above the study list (Add
Studies, Outcomes, Assign Reviewers) so the All Studies tab is just the
study list:

- Add studies: primary button in the sticky project header opening
  AddStudiesForm in a side sheet. The sheet owns the useAddStudies
  instance and stays mounted while closed, so staged studies survive
  close/reopen, page-level PDF drops stage files and open the sheet,
  and Google Drive OAuth restore reopens it with saved state.
- Assign reviewers: owner-only header button with an unassigned-count
  badge opening the bulk distribution flow in a sheet. Component moved
  from overview-tab/ to assign-reviewers/ and stripped of its card
  chrome. Per-study assignment from study cards is unchanged.
- Outcomes: header metadata line ("N studies . N outcomes") opens
  OutcomeManager in a sheet.
- After an import, owners get a toast with an action that opens the
  assign sheet for the newly unassigned studies.
- New ui/sheet.tsx primitive (radix dialog based, light blur overlay).
- Empty projects keep the inline add form as the centerpiece; the
  form's collapsible-card mode and document-level drag listeners moved
  to the sheet host.
